Fix whenfieldoriented/drive test case
The last test case of drive set method uses non-max speeds.
So the distance the module is away from the center of rotation of the robot will play a roll in the magnitude of the speed order.
The data being compared against in the test assumed a chassis size of 1x1"
Instead of implementing the dimension changes in our code, we're just overriding chassis dimensions within the test itself